Extract useLoader from useSearchLoader for reuse
